requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
feedparser>=6.0.11
openai>=1.50.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
